Decompile Overlay 3

+++ include/unk_02031B0C.h
@@ -19,7 +19,7 @@ int sub_02031D80(UnkStruct_02031CEC *a0);
 String *sub_020322AC(SaveApricornBox *apricornBox, u16 a1, u32 a2);
 String *sub_02032308(SaveApricornBox *apricornBox, u16 a1, u32 a2);
 int ApricornBox_CountApricorn(SaveApricornBox *apricornBox, u32 a1);
-void ApricornBox_GiveApricorn(SaveApricornBox *apricornBox, u16 a1, u8 a2);
+void ApricornBox_GiveApricorn(SaveApricornBox *apricornBox, u32 a1, u8 a2);
 int ApricornBox_GetKurtQuantity(SaveApricornBox *apricornBox);
 void ApricornBox_SetKurtApricorn(SaveApricornBox *apricornBox, u32 a0, u32 a1);
 int ApricornBox_GetKurtBall(SaveApricornBox *apricornBox);
